+++ features/overlays/rareCatchesTracker.js
@@ -4,9 +4,9 @@ import * as seaCreatures from '../../constants/seaCreatures';
 import { persistentData } from "../../data/data";
 import { overlayCoordsData } from "../../data/overlayCoords";
 import { formatNumberWithSpaces, fromUppercaseToCapitalizedFirstLetters, isDoubleHook, isInFishingWorld, pluralize } from '../../utils/common';
-import { WHITE, GOLD, BOLD, YELLOW, GRAY, RED } from "../../constants/formatting";
+import { WHITE, GOLD, BOLD, GRAY, RED, AQUA } from "../../constants/formatting";
 import { RARE_CATCH_TRIGGERS } from "../../constants/triggers";
-import { getWorldName, hasFishingRodInHotbar, isInSkyblock } from "../../utils/playerState";
+import { getLastFishingHookSeenAt, getWorldName, isInSkyblock } from "../../utils/playerState";
 import { createButtonsDisplay, toggleButtonsDisplay } from "../../utils/overlays";
 import { registerIf } from "../../utils/registers";
 
@@ -37,7 +37,7 @@ export function resetRareCatchesTracker(isConfirmed) {
     try {
         if (!isConfirmed) {
             new Message(
-                new TextComponent(`${GOLD}[FeeshNotifier] ${WHITE}Do you want to reset rare catches tracker? ${RED}${BOLD}[Click to confirm]`)
+                new TextComponent(`${GOLD}[FeeshNotifier] ${WHITE}Do you want to reset Rare catches tracker? ${RED}${BOLD}[Click to confirm]`)
                     .setClickAction('run_command')
                     .setClickValue('/feeshResetRareCatches noconfirm')
             ).chat();
@@ -77,7 +77,7 @@ function trackCatch(options) {
             return;
         }
     
-        if (options.seaCreature === seaCreatures.VANQUISHER && !hasFishingRodInHotbar()) {
+        if (options.seaCreature === seaCreatures.VANQUISHER && (new Date() - getLastFishingHookSeenAt() > 6 * 60 * 1000)) {
             return;
         }
 
@@ -118,14 +118,14 @@ function renderRareCatchTrackerOverlay() {
         !Object.entries(persistentData.rareCatches).length ||
         !isInSkyblock() ||
         !isInFishingWorld(getWorldName()) ||
-        !hasFishingRodInHotbar() ||
+        (new Date() - getLastFishingHookSeenAt() > 10 * 60 * 1000) ||
         allOverlaysGui.isOpen()
     ) {
         buttonsDisplay.hide();
         return;
     }
 
-    let overlayText = `${YELLOW}${BOLD}Rare catches tracker\n`;
+    let overlayText = `${AQUA}${BOLD}Rare catches tracker\n`;
 
     const entries = Object.entries(persistentData.rareCatches)
         .map(([key, value]) => {
@@ -142,7 +142,7 @@ function renderRareCatchTrackerOverlay() {
         overlayText += `${GRAY}- ${rarityColorCode}${fromUppercaseToCapitalizedFirstLetters(entry.seaCreature)}${GRAY}: ${WHITE}${formatNumberWithSpaces(entry.amount)}${doubleHookInfo}\n`;
     });
 
-    overlayText += `${YELLOW}Total: ${WHITE}${persistentData.totalRareCatches}`;
+    overlayText += `${GRAY}Total: ${WHITE}${persistentData.totalRareCatches}`;
 
     const overlay = new Text(overlayText, overlayCoordsData.rareCatchesTrackerOverlay.x, overlayCoordsData.rareCatchesTrackerOverlay.y)
         .setShadow(true)